+++ components/resources/quizzes/shared/SkillSelections.tsx
@@ -1,26 +1,11 @@
 import React from "react";
-import { QuizOption } from "../../../resources/quizzes/shared/types";
+import {
+  QuizOptionViewState,
+  QuizViewState,
+} from "../../../resources/quizzes/shared/types";
 import { Button } from "../../../ui/Button";
 import Progress from "./Progress";
 import SkillifyNavbar from "./SkillifyNavbar";
-
-export type QuizViewState = {
-  title: string;
-  body: string;
-  questions: QuizQuestionViewState[];
-  currentQuestion: number;
-  progress: number;
-};
-
-export type QuizQuestionViewState = {
-  lang: string;
-  title: string;
-  body: string;
-  options: QuizOptionViewState[];
-};
-export type QuizOptionViewState = QuizOption & {
-  isSelected: boolean;
-};
 
 type SkillSelectionsProps = {
   quizViewState: QuizViewState;
