+++ components/resources/quizzes/shared/types.ts
@@ -1,18 +1,39 @@
+//Types for static Data Structure
+export type QuizData = {
+  title: string;
+  body: string;
+  questions: QuizQuestion[];
+};
+
 export type QuizQuestion = {
     title: string;
     body: string;
     options: QuizOption[];
     maxSelections?: number;
   };
-  
-  export type QuizOption = {
+
+export type QuizOption = {
     name: string;
     result?: string;
     weight?: number;
   };
   
-  export type QuizData = {
+
+//Types for ViewState
+export type QuizViewState = {
     title: string;
     body: string;
-    questions: QuizQuestion[];
+    questions: QuizQuestionViewState[];
+    currentQuestion: number;
+    progress: number;
+  };
+  
+export type QuizQuestionViewState = {
+    title: string;
+    body: string;
+    options: QuizOptionViewState[];
+  };
+  
+export type QuizOptionViewState = QuizOption & {
+    isSelected: boolean;
   };
